factorize naming into Utils::NamedThorActions
