+++ lib/inferno/apps/cli/templates/README.md.tt
@@ -14,7 +14,7 @@ for FHIR testing.
 - Run `setup.sh` in this repo to pull the needed docker images and set up the
   database.
 - Put the `package.tgz` for the IG you're writing tests for in
-  `lib/<%= lib_name %>/igs`.
+  `lib/<%= library_name %>/igs`.
   This will ensure that the validator has access to the resources needed to
   validate resources against your IG.
 
@@ -108,7 +108,7 @@ end
 ᐅ bundle exec rdbg -A
 DEBUGGER (client): Connected. PID:22112, $0:sidekiq 6.5.7  [0 of 10 busy]
 
-[18, 27] in ~/code/<%= root_name %>/lib/<%= lib_name %>/patient_group.rb
+[18, 27] in ~/code/<%= root_name %>/lib/<%= library_name %>/patient_group.rb
     18|
     19|       run do
     20|         fhir_read(:patient, patient_id, name: :patient)
@@ -157,8 +157,8 @@ like a standard ruby gem (ruby libraries are called gems).
 - Fill in the information in the `gemspec` file in the root of this repository.
   The name of this file should match the `spec.name` within the file. This will
   be the name of the gem you create. For example, if your file is
-  `<%= lib_name %>.gemspec` and its `spec.name` is `'<%= lib_name %>'`, then others will
-  be able to install your gem with `gem install <%= lib_name %>`. There are
+  `<%= library_name %>.gemspec` and its `spec.name` is `'<%= library_name %>'`, then others will
+  be able to install your gem with `gem install <%= library_name %>`. There are
   [recommended naming conventions for
   gems](https://guides.rubygems.org/name-your-gem/).
 - Your tests must be in `lib`
