+++ lib/inferno/apps/cli/templates/lib/%library_name%.rb.tt
@@ -1,4 +1,4 @@
-require_relative '<%= lib_name %>/patient_group'
+require_relative '<%= library_name %>/patient_group'
 
 module <%= module_name %>
   class Suite < Inferno::TestSuite
