+++ lib/inferno/apps/cli/templates/lib/%library_name%/patient_group.rb.tt
